% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{latlong2eastnorth}
\alias{latlong2eastnorth}
\title{Latitude and Longitude to Easting and Northing}
\usage{
latlong2eastnorth(data, long = "Longitude", lat = "Latitude",
  east = "Easting", north = "Northing", geodatum = "WGS84",
  projargs = "+init=epsg:3005")
}
\arguments{
\item{data}{The data frame or spatial object with the locational information.}

\item{long}{A string of the name of the column in data with the longitudes.}

\item{lat}{A string of the name of the column in data with the latitudes.}

\item{east}{A string of the name of the column to add to data with the eastings.}

\item{north}{A string of the name of the column to add to data with the northings.}

\item{geodatum}{A string of the geodatum for the longitudes and latitudes.}

\item{projargs}{A string of the projection for the eastings and northings.}
}
\description{
By default convert latitudes and longitudes in the WGS84 geodatum to eastings and northings
in BC Albers projection.
}
