% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{latlong2kml}
\alias{latlong2kml}
\title{Latitude and Longitude to kml File.}
\usage{
latlong2kml(data, file = "plot.kml", label = "Site", long = "Longitude",
  lat = "Latitude", geodatum = "WGS84")
}
\arguments{
\item{data}{The data frame or spatial object with the locational information.}

\item{file}{A string specifying the file name.}

\item{label}{A string specifying the column with the labels.}

\item{long}{A string of the name of the column in data with the longitudes.}

\item{lat}{A string of the name of the column in data with the latitudes.}

\item{geodatum}{A string of the geodatum for the longitudes and latitudes.}
}
\description{
Latitude and Longitude to kml File.
}
